% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/util.R
\name{gen_map}
\alias{gen_map}
\title{generate a mice genetic map with one gape region on CHR}
\usage{
gen_map(n.mar.cM, gap.chr = 1, gap.start = 40, gap.end = 60,
  anchor.tel = TRUE, include.x = FALSE, sex.sp = FALSE,
  eq.spacing = FALSE)
}
\arguments{
\item{n.mar.cM}{Number of markers per cM}
}
\description{
generate a mice genetic map with one gape region on CHR
}

