% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gen_pheno_ail.R
\name{gen_pheno_ail}
\alias{gen_pheno_ail}
\title{generate pheno for ail data}
\usage{
gen_pheno_ail(xodat, map, id, qtl.chr, qtl.pos, h.qtl = 0.3, h.kin = 0.1,
  n.mqtl = 10)
}
\description{
generate pheno for ail data
}

