% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/gen_pheno_do.R
\name{gen_pheno_do}
\alias{gen_pheno_do}
\title{generate pheno from pedigree and xo data.}
\usage{
gen_pheno_do(xodat, map, id, qtl.chr, qtl.pos, allele, h.qtl = 0.3,
  h.kin = 0.1, n.mqtl = 10)
}
\description{
Add main qtl effect for a fixed qtl and multiple small effect QTLs
randomly selected from chr 2-19.
}

