% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gen_pheno_do.R
\name{gen_pheno_do_allqtl}
\alias{gen_pheno_do_allqtl}
\title{generate pheno for all the qtlpositons}
\usage{
gen_pheno_do_allqtl(xodat, map, id, qtl.chr, qtl.allpos, allele, h.qtl = 0.3,
  h.kin = 0.1, n.mqtl = 10, n.simu = 1)
}
\description{
returns a matrix of size n x \code{n_qtl*n.simu}
}

