% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/lod.drop.R
\name{lod.drop}
\alias{lod.drop}
\title{LOD-drop for lod support interval to reach {95\%} coverage rate.}
\usage{
lod.drop(pos, snp, LOD, qtl.pos, probs = 0.95, lod.thr = 3, eps = 1e-08)
}
\description{
LOD-drop for lod support interval to reach {95\%} coverage rate.
}

