% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/prep_data_founders.R
\name{prep_data_founders_chr}
\alias{prep_data_founders_chr}
\title{prepare for DOQTL::calc.genoprob}
\usage{
prep_data_founders_chr(ped, xodat, map, id, f.geno, chr)
}
\description{
generate list 'data' and 'founders' from simulated DO pedigree and
crossovers for a chromosome.
}

