% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/qtlsummary.do.R
\name{qtlsummary.do}
\alias{qtlsummary.do}
\title{summarize of do simulation results}
\usage{
qtlsummary.do(LOD, pos, snp, qtl.pos, lod.thr = 3, drop = 1.5, out.qtl)
}
\description{
take LOD score and marker infomation, return coverage and interval
width for LOD support interval, Bayes interval, lod.drop and so on.
}

