% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/run.do.single.R
\name{run.do.single}
\alias{run.do.single}
\title{run simulations for DO type experiments}
\usage{
run.do.single(method = c("sub2", "last2", "fixcc"), para, i.para, i.simu,
  i.qtlpos, output.dir = "DO.output/", result.dir = "./",
  n.simu_qtlpos = 1, n.simu = 10000, qtl.chr = 1, ochr = 2:19,
  n.mqtl = 10, n.ccgen = 15, map.whole, qtl.allpos, f.geno.chr, snps,
  design = c("nosib", "random"), selc.method = c("byfamily", "byindiv"),
  write.gp36 = FALSE, cleanup = TRUE, realpb8 = FALSE, noK = FALSE)
}
\arguments{
\item{method}{design method for the pedigree, choose from 'sub2' or
'last2' or 'fixcc'.}

\item{para}{data.frame that saves all the parameter settings for the simulation}

\item{i.para}{which setting to use for the current run}

\item{i.simu}{which simulation is the currtent run at}

\item{i.qtlpos}{which candidate qtlpos will be used.}

\item{n.simu_qtlpos}{number of phenotypes for each qtl position. suggested
to be 1, otherwise the result will be correlated.}

\item{qtl.chr}{chromosome for the qtl}

\item{ochr}{chromosomes other than qtl.chr, used to estimate kinship matrix}

\item{n.mqtl}{number of minor qtls for the polygenic effects}

\item{map.whole}{genetic map for all chromosomes}

\item{qtl.allpos}{named vector of all candidate qtl positions.}

\item{f.geno.chr}{founder genotype for QTL chromosome.}

\item{snps}{SNP information saved in a data.frame: name, chr and
position. Only for SNPs on the QTL chromosome.}

\item{design}{'nosib' or 'random'}

\item{write.gp36}{logic value}

\item{cleanup}{logic value}

\item{realpb8}{will the real 8 allele probabilities be used while
mapping.}

\item{noK}{do we control for Kinship matrix while mapping.}
}
\description{
For a fix set of parameters, generate DO pedigree and simulate
crossovers, randomly generate phenotypes with main QTL and minor
QTLs, then run DOQTL::scanone to do QTL mapping. save LOD score as
a matrix and marker informations to result.dir.
}

