% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/util.R
\name{addblank}
\alias{addblank}
\title{add a gap region into a chr by removing markers.}
\usage{
addblank(map, chr, gap.start, gap.end)
}
\description{
add a gap region into a chr by removing markers.
}

