% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/gen_pheno_do.R
\name{get_minor_qtl_do}
\alias{get_minor_qtl_do}
\title{generate some minor qtls from other Chrs, not X and the QTL chr.}
\usage{
get_minor_qtl_do(xodat, id, map, n.mqtl, qtl.chr)
}
\description{
generate some minor qtls from other Chrs, not X and the QTL chr.
}

