% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/prep_data_founders.R
\name{prep_data_founders}
\alias{prep_data_founders}
\title{prepare for DOQTL::calc.genoprob}
\usage{
prep_data_founders(ped, xodat, map, id, f.geno)
}
\description{
generate list 'data' and 'founders' from simulated DO pedigree and
crossovers.
}

