% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/run.ail.R
\name{run.ail}
\alias{run.ail}
\title{run simulations for AIL}
\usage{
run.ail(method = c("sub2", "last2"), para, i.para, n.simu,
  result.dir = "./", qtl.chr = 1, ochr = 2:19, n.mqtl = 10,
  design = c("nosib", "random"))
}
\arguments{
\item{method}{design method for the pedigree, choose from 'sub2' or 'last2'.}

\item{para}{data.frame that saves all the parameter settings for the simulation}

\item{i.para}{which setting to use for the current run}

\item{n.simu}{number of phenotypes for the current run}

\item{qtl.chr}{chromosome for the qtl}

\item{ochr}{chromosomes other than qtl.chr, used to estimate kinship matrix}

\item{n.mqtl}{number of minor qtls for the polygenic effects}

\item{design}{'nosib' or 'random'}
}
\description{
For a fix set of parameters, generate AIL pedigree and simulate
crossovers, randomly generate phenotypes with main QTL and minor
QTLs, then run QTLRel::scanOne to do QTL mapping. save LOD score as
a matrix and marker infomations to result.dir.
}

