% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/run.do.R
\name{run.do}
\alias{run.do}
\title{run simulations for DO type experiments}
\usage{
run.do(method = "sub2", para, i.para, j.simu, output.dir = "DO.output/",
  result.dir = ".", n.simu = 1, qtl.chr = 1, ochr = 2:19, n.mqtl = 10,
  map.whole, qtl.allpos, f.geno.chr, snps, write.gp36 = FALSE,
  cleanup = TRUE)
}
\arguments{
\item{n.simu}{number of phenotypes for each qtl position. suggested
to be 1, otherwise the result will be correlated.}
}
\description{
For a fix set of parameters, generate DO pedigree and simulate
crossovers, randomly generate phenotypes with main QTL and minor
QTLs, then run DOQTL::scanone to do QTL mapping. save LOD score as
a matrix and marker infomations to result.dir.
}

